requirements.txt
# Adaptive Retail Demand Forecasting System
# Core dependencies

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Drift Detection
river>=0.21.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0

# Web Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Calendar/Holiday features
holidays>=0.36
